Move the node createWorker() method to the NodeMixin.
